+++ wiring/src/spark_wiring_system.cpp
@@ -35,7 +35,7 @@ void SystemClass::sleep(Spark_Sleep_TypeDef sleepMode, long seconds)
     system_sleep(sleepMode, seconds, 0, NULL);
 }
 
-void SystemClass::sleep(uint16_t wakeUpPin, uint16_t edgeTriggerMode, long seconds)
+void SystemClass::sleep(uint16_t wakeUpPin, InterruptMode edgeTriggerMode, long seconds)
 {
     system_sleep_pin(wakeUpPin, edgeTriggerMode, seconds, 0, NULL);
 }