+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### ENHANCEMENTS
  - Retrieve the LED brightness via `RGB.brightness()`
+ - System.sleep() - 2nd parameter changed to `InterruptMode` from uint16_t to
+ ensure the correct types are used. [#499](https://github.com/spark/firmware/pull/499)
 
 ### BUGFIXES